+++ pyvrp/cpp/ProblemData.h
@@ -1,14 +1,15 @@
 #ifndef PYVRP_PROBLEMDATA_H
 #define PYVRP_PROBLEMDATA_H
 
+#include "Client.h"
+#include "ClientGroup.h"
+#include "Depot.h"
+#include "Location.h"
 #include "Matrix.h"
 #include "Measure.h"
+#include "VehicleType.h"
 
 #include <cassert>
-#include <iosfwd>
-#include <limits>
-#include <optional>
-#include <string>
 #include <vector>
 
 namespace pyvrp
@@ -63,543 +64,6 @@ namespace pyvrp
  */
 class ProblemData
 {
-    // Validates the consistency of the constructed instance.
-    void validate() const;
-
-public:
-    /**
-     * Location(
-     *     x: float,
-     *     y: float,
-     *     *,
-     *     name: str = "",
-     * )
-     *
-     * Represents a physical location.
-     *
-     * Parameters
-     * ----------
-     * x
-     *     Horizontal coordinate of this location. This can for example be a
-     *     longitude value.
-     * y
-     *     Vertical coordinate of this location. This can for example be a
-     *     latitude value.
-     * name
-     *     Free-form name field for this location. Default empty.
-     *
-     * Attributes
-     * ----------
-     * x
-     *     Horizontal location coordinate.
-     * y
-     *     Vertical location coordinate.
-     * name
-     *     Free-form name field for this location.
-     */
-    struct Location
-    {
-        Coordinate const x;
-        Coordinate const y;
-        char const *name;
-
-        Location(Coordinate x, Coordinate y, std::string name = "");
-
-        bool operator==(Location const &other) const;
-
-        Location(Location const &location);
-        Location(Location &&location);
-
-        Location &operator=(Location const &location) = delete;
-        Location &operator=(Location &&location) = delete;
-
-        ~Location();
-    };
-
-    /**
-     * Client(
-     *    location: int,
-     *    delivery: list[int] = [],
-     *    pickup: list[int] = [],
-     *    service_duration: int = 0,
-     *    tw_early: int = 0,
-     *    tw_late: int = np.iinfo(np.int64).max,
-     *    release_time: int = 0,
-     *    prize: int = 0,
-     *    required: bool = True,
-     *    group: int | None = None,
-     *    *,
-     *    name: str = "",
-     * )
-     *
-     * Simple data object storing all client data as properties. See also
-     * :doc:`../setup/concepts` for further information about these properties.
-     *
-     * Parameters
-     * ----------
-     * location
-     *     Physical location of this client.
-     * delivery
-     *     The amounts this client demands from the depot.
-     * pickup
-     *     The amounts this client ships back to the depot.
-     * service_duration
-     *     Amount of time a vehicle needs to spend at this client before
-     *     resuming its route. Service should start (but not necessarily end)
-     *     within the [:py:attr:`~tw_early`, :py:attr:`~tw_late`] interval.
-     *     Default 0.
-     * tw_early
-     *     Earliest time at which this client may be visited to start service.
-     *     Default 0.
-     * tw_late
-     *     Latest time at which this client may be visited to start service.
-     *     Unconstrained if not provided.
-     * release_time
-     *     Earliest time at which this client is released, that is, the earliest
-     *     time at which a vehicle may leave the depot on a trip to visit this
-     *     client. Default 0.
-     * prize
-     *     Prize collected by visiting this client. Default 0. If this client
-     *     is not required, the prize needs to be sufficiently large to offset
-     *     any travel cost before this client will be visited in a solution.
-     * required
-     *     Whether this client must be part of a feasible solution. Default
-     *     True. Make sure to also update the prize value when setting this
-     *     argument to False.
-     * group
-     *     Indicates membership of the given client group, if any. By default
-     *     clients are not part of any groups.
-     * name
-     *     Free-form name field for this client. Default empty.
-     *
-     * Attributes
-     * ----------
-     * location
-     *     Physical location of this client.
-     * delivery
-     *     Client delivery amounts shipped from the depot.
-     * pickup
-     *     Client pickup amounts returned to the depot.
-     * service_duration
-     *     Amount of time a vehicle needs to spend at this client before
-     *     resuming its route.
-     * tw_early
-     *     Earliest time at which this client may be visited to start service.
-     * tw_late
-     *     Latest time at which this client may be visited to start service.
-     * release_time
-     *     Earliest time at which a vehicle may leave the depot on a trip to
-     *     visit this client.
-     * prize
-     *     Prize collected by visiting this client.
-     * required
-     *     Whether visiting this client is required.
-     * group
-     *     Indicates membership of the given client group, if any.
-     * name
-     *     Free-form name field for this client.
-     */
-    struct Client
-    {
-        size_t const location;
-        Duration const serviceDuration;
-        Duration const twEarly;  // Earliest possible start of service
-        Duration const twLate;   // Latest possible start of service
-        std::vector<Load> const delivery;
-        std::vector<Load> const pickup;
-        Duration const releaseTime;  // Earliest possible time to leave depot
-        Cost const prize;            // Prize for visiting this client
-        bool const required;         // Must client be in solution?
-        std::optional<size_t> const group;  // Optional client group membership
-        char const *name;                   // Client name (for reference)
-
-        Client(size_t location,
-               std::vector<Load> delivery = {},
-               std::vector<Load> pickup = {},
-               Duration serviceDuration = 0,
-               Duration twEarly = 0,
-               Duration twLate = std::numeric_limits<Duration>::max(),
-               Duration releaseTime = 0,
-               Cost prize = 0,
-               bool required = true,
-               std::optional<size_t> group = std::nullopt,
-               std::string name = "");
-
-        bool operator==(Client const &other) const;
-
-        Client(Client const &client);
-        Client(Client &&client);
-
-        Client &operator=(Client const &client) = delete;
-        Client &operator=(Client &&client) = delete;
-
-        ~Client();
-    };
-
-    /**
-     * ClientGroup(
-     *    clients: list[int] = [],
-     *    required: bool = True,
-     *    *,
-     *    name: str = "",
-     * )
-     *
-     * A client group that imposes additional restrictions on visits to clients
-     * in the group.
-     *
-     * .. note::
-     *
-     *    Only mutually exclusive client groups are supported for now.
-     *
-     * Parameters
-     * ----------
-     * clients
-     *     The clients in the group.
-     * required
-     *     Whether visiting this client group is required.
-     * name
-     *    Free-form name field for this client group. Default empty.
-     *
-     * Attributes
-     * ----------
-     * clients
-     *     The clients in the group.
-     * required
-     *     Whether visiting this client group is required.
-     * mutually_exclusive
-     *     When ``True``, exactly one of the clients in this group must be
-     *     visited if the group is required, and at most one if the group is
-     *     not required.
-     * name
-     *    Free-form name field for this client group.
-     *
-     * Raises
-     * ------
-     * ValueError
-     *     When the given clients contain duplicates, or when a client is added
-     *     to the group twice.
-     */
-    class ClientGroup
-    {
-        std::vector<size_t> clients_;  // clients in this group
-
-    public:
-        bool const required;                  // is visiting the group required?
-        bool const mutuallyExclusive = true;  // at most one visit in group?
-        char const *name;                     // Group name (for reference)
-
-        explicit ClientGroup(std::vector<size_t> clients = {},
-                             bool required = true,
-                             std::string name = "");
-
-        bool operator==(ClientGroup const &other) const;
-
-        ClientGroup(ClientGroup const &group);
-        ClientGroup(ClientGroup &&group);
-
-        ClientGroup &operator=(ClientGroup const &group) = delete;
-        ClientGroup &operator=(ClientGroup &&group) = delete;
-
-        ~ClientGroup();
-
-        bool empty() const;
-        size_t size() const;
-
-        std::vector<size_t>::const_iterator begin() const;
-        std::vector<size_t>::const_iterator end() const;
-
-        std::vector<size_t> const &clients() const;
-
-        void addClient(size_t client);
-    };
-
-    /**
-     * Depot(
-     *    location: int,
-     *    tw_early: int = 0,
-     *    tw_late: int = np.iinfo(np.int64).max,
-     *    service_duration: int = 0,
-     *    *,
-     *    name: str = "",
-     * )
-     *
-     * Simple data object storing all depot data as (read-only) properties.
-     *
-     * Parameters
-     * ----------
-     * location
-     *     Physical location of this depot.
-     * tw_early
-     *     Opening time of this depot. Default 0.
-     * tw_late
-     *     Closing time of this depot. Default unconstrained.
-     * service_duration
-     *     Time it takes to e.g. load a vehicle at this depot, at the start of
-     *     a trip. Default 0.
-     * name
-     *     Free-form name field for this depot. Default empty.
-     *
-     * Attributes
-     * ----------
-     * location
-     *     Physical location of this depot.
-     * tw_early
-     *     Opening time of this depot.
-     * tw_late
-     *     Closing time of this depot.
-     * service_duration
-     *     Time it takes to e.g. load a vehicle at this depot, at the start of
-     *     a trip.
-     * name
-     *     Free-form name field for this depot.
-     */
-    struct Depot
-    {
-        size_t const location;
-        Duration const serviceDuration;
-        Duration const twEarly;  // Depot opening time
-        Duration const twLate;   // Depot closing time
-        char const *name;        // Depot name (for reference)
-
-        Depot(size_t location,
-              Duration twEarly = 0,
-              Duration twLate = std::numeric_limits<Duration>::max(),
-              Duration serviceDuration = 0,
-              std::string name = "");
-
-        bool operator==(Depot const &other) const;
-
-        Depot(Depot const &depot);
-        Depot(Depot &&depot);
-
-        Depot &operator=(Depot const &depot) = delete;
-        Depot &operator=(Depot &&depot) = delete;
-
-        ~Depot();
-    };
-
-    /**
-     * VehicleType(
-     *     num_available: int = 1,
-     *     capacity: list[int] = [],
-     *     start_depot: int = 0,
-     *     end_depot: int = 0,
-     *     fixed_cost: int = 0,
-     *     tw_early: int = 0,
-     *     tw_late: int = np.iinfo(np.int64).max,
-     *     shift_duration: int = np.iinfo(np.int64).max,
-     *     max_distance: int = np.iinfo(np.int64).max,
-     *     unit_distance_cost: int = 1,
-     *     unit_duration_cost: int = 0,
-     *     profile: int = 0,
-     *     start_late: int | None = None,
-     *     initial_load: list[int] = [],
-     *     reload_depots: list[int] = [],
-     *     max_reloads: int = np.iinfo(np.uint64).max,
-     *     max_overtime: int = 0,
-     *     unit_overtime_cost: int = 0,
-     *     *,
-     *     name: str = "",
-     * )
-     *
-     * Simple data object storing all vehicle type data as properties. See also
-     * :doc:`../setup/concepts` for further information about these properties.
-     *
-     * Parameters
-     * ----------
-     * num_available
-     *     Number of vehicles of this type that are available. Must be positive.
-     *     Default 1.
-     * capacity
-     *     Capacities of this vehicle type, per load dimension. This capacity is
-     *     the maximum total delivery or pickup amount that the vehicle can
-     *     store along the route.
-     * start_depot
-     *     Depot (index) where vehicles of this type start their routes.
-     *     Defaults to the first depot.
-     * end_depot
-     *     Depot (index) where vehicles of this type end routes. Defaults to
-     *     the first depot.
-     * fixed_cost
-     *     Fixed cost of using a vehicle of this type. Default 0.
-     * tw_early
-     *     Start of the vehicle type's shift. Default 0.
-     * tw_late
-     *     End of the vehicle type's shift. Unconstrained if not provided.
-     * shift_duration
-     *     Nominal maximum route duration. May be extended through overtime
-     *     (see :py:attr:`~max_overtime`) at additional cost. Unconstrained if
-     *     not explicitly provided.
-     * max_distance
-     *     Maximum route distance. Unconstrained if not explicitly provided.
-     * unit_distance_cost
-     *     Cost per unit of distance travelled by vehicles of this type. Default
-     *     1.
-     * unit_duration_cost
-     *     Cost per unit of duration on routes serviced by vehicles of this
-     *     type. Default 0.
-     * profile
-     *     This vehicle type's routing profile. Default 0, the first profile.
-     * start_late
-     *     Latest start of the vehicle type's shift. Unconstrained if not
-     *     provided.
-     * initial_load
-     *     Load already on the vehicle that need to be dropped off at a depot.
-     *     This load is present irrespective of any client visits. By default
-     *     this value is zero, and the vehicle only considers loads from client
-     *     visits.
-     * reload_depots
-     *     List of reload depots (indices) this vehicle may visit along its
-     *     route, to empty and reload for subsequent client visits. Defaults to
-     *     an empty list, in which case no reloads are allowed.
-     * max_reloads
-     *     Maximum number of reloads the vehicle may perform on a route.
-     *     Unconstrained if not explicitly provided.
-     * max_overtime
-     *     Maximum allowed overtime, on top of the :py:attr:`~shift_duration`.
-     *     Default 0, that is, overtime is not allowed.
-     * unit_overtime_cost
-     *     Cost of a unit of overtime. This is in addition to the regular
-     *     :py:attr:`~unit_duration_cost` of route durations. Default 0.
-     * name
-     *     Free-form name field for this vehicle type. Default empty.
-     *
-     * Attributes
-     * ----------
-     * num_available
-     *     Number of vehicles of this type that are available.
-     * capacity
-     *     Capacities of this vehicle type, per load dimension.
-     * start_depot
-     *     Start depot associated with these vehicles.
-     * end_depot
-     *     End depot associated with these vehicles.
-     * fixed_cost
-     *     Fixed cost of using a vehicle of this type.
-     * tw_early
-     *     Start of the vehicle type's shift, if specified.
-     * tw_late
-     *     End of the vehicle type's shift, if specified.
-     * shift_duration
-     *     Nominal maximum shift duration of the route this vehicle type is
-     *     assigned to. Default unconstrained.
-     * max_distance
-     *     Maximum travel distance of the route this vehicle type is assigned
-     *     to. This is a very large number when the maximum distance is
-     *     unconstrained.
-     * unit_distance_cost
-     *     Cost per unit of distance travelled by vehicles of this type.
-     * unit_duration_cost
-     *     Cost per unit of duration on routes using vehicles of this type.
-     * profile
-     *     This vehicle type's routing profile.
-     * start_late
-     *     Latest start of the vehicle type's shift. This is equal to
-     *     ``tw_late`` when the latest start is not constrained.
-     * initial_load
-     *     Load already on the vehicle that need to be dropped off at a depot.
-     *     This load is present irrespective of any client visits.
-     * reload_depots
-     *     List of reload locations this vehicle may visit along it route, to
-     *     empty and reload.
-     * max_reloads
-     *     Maximum number of reloads the vehicle may perform on a route.
-     * max_overtime
-     *     Maximum amount of allowed overtime, on top of the nominal
-     *     :py:attr:`~shift_duration`.
-     * unit_overtime_cost
-     *     Additional cost of a unit of overtime.
-     * max_duration
-     *     Hard maximum route duration constraint, computed as the sum of
-     *     :py:attr:`~shift_duration` and :py:attr:`~max_overtime`.
-     * name
-     *     Free-form name field for this vehicle type.
-     */
-    struct VehicleType
-    {
-        size_t const numAvailable;         // Available vehicles of this type
-        size_t const startDepot;           // Departure depot location
-        size_t const endDepot;             // Return depot location
-        std::vector<Load> const capacity;  // This type's vehicle capacity
-        Duration const twEarly;            // Start of shift
-        Duration const twLate;             // End of shift
-        Duration const shiftDuration;      // Nominal shift duration
-        Distance const maxDistance;        // Maximum route distance
-        Cost const fixedCost;         // Fixed cost of using this vehicle type
-        Cost const unitDistanceCost;  // Variable cost per unit of distance
-        Cost const unitDurationCost;  // Variable cost per unit of duration
-        size_t const profile;         // Distance and duration profile
-        Duration const startLate;     // Latest start of shift
-        std::vector<Load> const initialLoad;     // Initially used capacity
-        std::vector<size_t> const reloadDepots;  // Reload locations
-        size_t const maxReloads;                 // Maximum number of reloads
-        Duration const maxOvertime;              // Maximum allowed overtime
-        Cost const unitOvertimeCost;             // Cost per unit of overtime
-        Duration const maxDuration;  // Maximum route duration, incl. overtime
-        char const *name;            // Type name (for reference)
-
-        VehicleType(size_t numAvailable = 1,
-                    std::vector<Load> capacity = {},
-                    size_t startDepot = 0,
-                    size_t endDepot = 0,
-                    Cost fixedCost = 0,
-                    Duration twEarly = 0,
-                    Duration twLate = std::numeric_limits<Duration>::max(),
-                    Duration shiftDuration
-                    = std::numeric_limits<Duration>::max(),
-                    Distance maxDistance = std::numeric_limits<Distance>::max(),
-                    Cost unitDistanceCost = 1,
-                    Cost unitDurationCost = 0,
-                    size_t profile = 0,
-                    std::optional<Duration> startLate = std::nullopt,
-                    std::vector<Load> initialLoad = {},
-                    std::vector<size_t> reloadDepots = {},
-                    size_t maxReloads = std::numeric_limits<size_t>::max(),
-                    Duration maxOvertime = 0,
-                    Cost unitOvertimeCost = 0,
-                    std::string name = "");
-
-        bool operator==(VehicleType const &other) const;
-
-        VehicleType(VehicleType const &vehicleType);
-        VehicleType(VehicleType &&vehicleType);
-
-        VehicleType &operator=(VehicleType const &vehicleType) = delete;
-        VehicleType &operator=(VehicleType &&vehicleType) = delete;
-
-        ~VehicleType();
-
-        /**
-         * Returns a new ``VehicleType`` with the same data as this one, except
-         * for the given parameters, which are used instead.
-         */
-        VehicleType replace(std::optional<size_t> numAvailable,
-                            std::optional<std::vector<Load>> capacity,
-                            std::optional<size_t> startDepot,
-                            std::optional<size_t> endDepot,
-                            std::optional<Cost> fixedCost,
-                            std::optional<Duration> twEarly,
-                            std::optional<Duration> twLate,
-                            std::optional<Duration> shiftDuration,
-                            std::optional<Distance> maxDistance,
-                            std::optional<Cost> unitDistanceCost,
-                            std::optional<Cost> unitDurationCost,
-                            std::optional<size_t> profile,
-                            std::optional<Duration> startLate,
-                            std::optional<std::vector<Load>> initialLoad,
-                            std::optional<std::vector<size_t>> reloadDepots,
-                            std::optional<size_t> maxReloads,
-                            std::optional<Duration> maxOvertime,
-                            std::optional<Cost> unitOvertimeCost,
-                            std::optional<std::string> name) const;
-
-        /**
-         * Returns the maximum number of trips these vehicle can execute.
-         */
-        size_t maxTrips() const;
-    };
-
-private:
     std::vector<Matrix<Distance>> const dists_;    // Distance matrices
     std::vector<Matrix<Duration>> const durs_;     // Duration matrices
     std::vector<Location> const locations_;        // Location information
@@ -611,6 +75,9 @@ private:
     size_t const numVehicles_;
     size_t const numLoadDimensions_;
     bool const hasTimeWindows_;
+
+    // Validates the consistency of the constructed instance.
+    void validate() const;
 
 public:
     bool operator==(ProblemData const &other) const = default;
@@ -840,13 +307,13 @@ public:
     ProblemData() = delete;
 };
 
-ProblemData::Client const &ProblemData::client(size_t client) const
+Client const &ProblemData::client(size_t client) const
 {
     assert(client < numClients());
     return clients_[client];
 }
 
-ProblemData::Depot const &ProblemData::depot(size_t depot) const
+Depot const &ProblemData::depot(size_t depot) const
 {
     assert(depot < numDepots());
     return depots_[depot];
